task4-fine-tuning/granulometry/requirements.txt
torch>=2.0
transformers>=4.45
peft>=0.12
trl>=0.9
bitsandbytes>=0.43
accelerate>=0.34
Pillow
anthropic
openai
